CATS-3586 | Cache tooltip responses for 5 minutes on the CDN
It seems that on the backend API side, there was already logic added
in 0f5a88e069102a6acfe2bb7e591113439a0605d3 to enable CDN caching for
response, but it was not working because the tooltip API calls were
using POST requests, which are ipso facto uncacheable.

The 'options' parameter for tooltip API calls can only ever take a
limited list of well-known options, so we should be safe to switch all
tooltip API calls to use GET requests without having to worry
about URL length. Unfortunately, the MediaWiki API does not have an
effective way of reusing cached responses for different registered
users, so we can only cache this response for anonymous users - but this
should be sufficient for the purpose of mitigating traffic spikes when
interest in a wiki using tooltips spikes.

As a further optimization, specify a background revalidation grace time
(stale-while-revalidate) of 1 minute in addition to the 5 minute cache
TTL to increase the chance of being able to revalidate a stale response,
and add conditional revalidation support to the API module based on the
modified (touched) timestamps of the tooltip page and the target page,
if applicable.