Avoid querying the database for non-redirect pages
When parsing large pages, TippingOver may end up triggering thousands of
DB queries from followRedirect(), because it creates a full WikiPage on
each invocation. Gate this behind an isRedirect() check, which can
leverage LinkCache to hopefully avoid a DB query in the common case that
the given title is not a redirect.